fix(cli): auto-adjust incompatible create flags with explanations
The non-interactive create flow never ran analyzeStackCompatibility, so
invalid flag combos were silently normalized (e.g. an unsupported flag
value coerced to a default with no message) or hard-failed, while the
web builder and MCP flows auto-adjust with explanations.

- Extract buildCompatibilityInputFromConfig / compatibilityChangesToProjectConfig
  from stack-update.ts into a shared utils/stack-compatibility.ts (no third
  mapper), plus a new resolveCompatibilityAdjustments() that diffs the engine
  result back into ProjectConfig keys.
- Run the pass in createProjectHandler on both non-interactive branches:
  the --yes/--part/config-file/--from-history branch (full config) and the
  flag path (provided flags only, before strict validation so combos like
  --frontend svelte --api trpc adjust to orpc instead of erroring).
- Print a user-facing summary per adjustment: `option: <from> -> <to> — <reason>`;
  stay silent when nothing changes.
- Respect --yolo (bypass) and skip programmatic/silent callers so the strict
  validation contract for the create() API, tests, and integrations is unchanged.
- Map config.effect through the engine's backendLibraries field (was hardcoded
  "none", causing spurious effect-backend adjustments) and give the react-native
  --yes base config database/orm "none" so default RN runs stay adjustment-free.
